requirements.txt
glfw~=2.5.3
pyopengl~=3.1.6
scapy~=2.4.5
pyglm
pywin32
numpy
